requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
peft==0.6.2
timm==0.9.7
faiss-cpu==1.7.4
fastapi==0.104.0
uvicorn[standard]==0.24.0
httpx==0.25.0
cryptography==41.0.5
pillow==10.0.1
numpy==1.26.0
pyyaml==6.0.1
prometheus-client==0.18.0
scipy==1.11.3
scikit-learn==1.3.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.10.1
flake8==6.1.0
mypy==1.6.1
psutil==5.9.6
opencv-python==4.8.1.78